Option to load crazyflies.yaml from string or alternate path. (#189)
Option to load launch YAML from string or alternate path.

Until now, the set of robots in a pycrazyswarm script is always
determined by ros_ws/src/crazyswarm/launch/crazyflies.yaml. This
couples the result of the script to the state of the executing machine.
If we want to add more automated testing in the future, it should be
possible to write a pycrazyswarm script that runs deterministically
without the cumbersome process of writing .yaml to disk.