Add files via upload
We modify and add two scrips that will let us simulate crazyswarm's code in gazebo based in the CrazyS repo.

crazyswarm_py.py
This file has a new parser.argument that let us use the --gaz in terminal while running your python's trajectory script to launch drones in Gazebo. Before doing that, remember to run the .yaml file to select the number of drones that will be used in the fly and  the .launch file in terminal to create the drone's world.

crazyflieGaz.py
This file contains the three fundamental methods for making your drones fly (Take off, go to and land). The above through nodes and topics which will be used by Gazebo while running the trajectory in your python script. Don't forget to check whether your script has all the arguments that the methods in this file require.