server(cpp): disable query of parameter values by default
Querying all 200+ parameters increases the connection time significantly. At the same time, the actual values are often not needed. Here, we disable the query by default, but the user can get the old behavior by changing a (global) setting. In the future, it might be useful to have a service call to update a list of given parameters (e.g., to query of some deck was initialized).